Switch to the container-based infrastructure on Travis
This uses the PHP builtin webserver to expose web fixtures rather than
installing Apache. This relies on the fact that PHP 5.6 is installed on
Travis even on jobs running with a different version, and so can be used
for fixtures on all jobs (and we don't care about the PHP version used
to run fixtures as we were not using the Travis one before anyway).